requirements.txt
# Python 3.13 Compatible Requirements
python-telegram-bot>=21.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
ta>=0.10.2
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
matplotlib>=3.8.0
seaborn>=0.13.0
schedule>=1.2.0
apscheduler>=3.10.4

# Optional packages (may not work with Python 3.13)
# pillow>=10.1.0
# plotly>=5.17.0
# websocket-client>=1.7.0